Redesign swift package build
Build first schema instead of all targets. Why first schema? Because
if there is no xcodeproj, it will be generated via `swift package
generate-xcodeproj` and <PACKAGE_NAME>-Package schema will be
generated by `swift package` -- this is what should be built. In other
case, when there is xcodeproj, the first schema will be built because
seems like Carthage did this