draft GitRemoteCacher
Logic is simillar to GitCacher but a little bit different -- need to
refactor, but idea is nice. Work only with remote repo and clone it
with --depth=1 and specified branch name, which optimize the size of
repo (don't download everythin, which it do, from the web). Also the
organization of repo is well defined right now -- repo, where each
build has unique branch. In terms of remote git repo (github etc) it
will perfectly work, and will work with cis also