add Cacher.update because updating in save isn't optimal
When saving you should clone master branch (because your package
branch isn't created yes -- it's save), but when updating you can
directly clone your package branch (without cloning anything else) and
work with that.

In future this should be simplified from user's point of view, but
right now it is as it is.