return [String] from PodManager.move
This return paths of destination builds which is great, but the main
problem why caching build for PodManager is pain is building all pods
and then precessing all pods. Building some pod may produce multiple
outputs and it cannot be deduced in processing stage. The solution is
to redesign PodManager as SPMManager -- build pod, merge pod and do it
for all of them